refactor decision table pages to leverage new python tools for decision table presentation
